beam_pipeline/requirements.txt
apache-beam==2.40.0
cachetools==4.2.4
certifi==2022.6.15
charset-normalizer==2.1.1
cloudpickle==2.1.0
crcmod==1.7
dill==0.3.1.1
docopt==0.6.2
fastavro==1.6.0
fasteners==0.17.3
google-api-core==1.32.0
google-apitools==0.5.31
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-cloud-bigquery==2.34.4
google-cloud-bigquery-storage==2.14.2
google-cloud-bigtable==1.7.2
google-cloud-core==1.7.3
google-cloud-datastore==1.15.5
google-cloud-dlp==3.8.1
google-cloud-language==1.3.2
google-cloud-pubsub==2.13.6
google-cloud-pubsublite==1.4.3
google-cloud-recommendations-ai==0.2.0
google-cloud-spanner==1.19.3
google-cloud-translate==3.8.1
google-cloud-videointelligence==1.16.3
google-cloud-vision==1.0.2
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
grpc-google-iam-v1==0.12.4
grpcio==1.47.0
grpcio-gcp==0.2.2
grpcio-status==1.47.0
hdfs==2.7.0
httplib2==0.20.4
idna==3.3
numpy==1.22.4
oauth2client==4.1.3
orjson==3.7.12
overrides==6.2.0
packaging==21.3
proto-plus==1.22.0
protobuf==3.20.1
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydot==1.4.2
pymongo==3.12.3
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
rsa==4.9
six==1.16.0
typing_extensions==4.3.0
urllib3==1.26.11
